# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dokumentasi Main API",
        "description": "Tulang punggung dari berbagai aplikasi.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/motorinci/specification-items": {
            "get": {
                "tags": [
                    "Motorinci Specification Items"
                ],
                "summary": "Mendapatkan daftar item spesifikasi motorinci",
                "operationId": "getMotorinciSpecificationItemsList",
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/200_Success"
                    },
                    "401": {
                        "$ref": "#/components/responses/401_Unauthorized"
                    }
                }
            },
            "post": {
                "tags": [
                    "Motorinci Specification Items"
                ],
                "summary": "Membuat item spesifikasi motorinci baru",
                "operationId": "storeMotorinciSpecificationItem",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "specification_group_id",
                                    "name"
                                ],
                                "properties": {
                                    "specification_group_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Panjang x Lebar x Tinggi"
                                    },
                                    "unit": {
                                        "type": "string",
                                        "example": "mm"
                                    },
                                    "desc": {
                                        "type": "string",
                                        "example": "Ukuran dimensi total motor."
                                    },
                                    "order": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "$ref": "#/components/responses/201_Created"
                    },
                    "422": {
                        "$ref": "#/components/responses/422_UnprocessableContent"
                    },
                    "401": {
                        "$ref": "#/components/responses/401_Unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/403_Forbidden"
                    }
                }
            }
        },
        "/api/motorinci/specification-items/{id}": {
            "get": {
                "tags": [
                    "Motorinci Specification Items"
                ],
                "summary": "Mendapatkan satu item spesifikasi motorinci berdasarkan ID",
                "operationId": "getMotorinciSpecificationItemById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Motorinci specification item ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/200_Success"
                    },
                    "404": {
                        "$ref": "#/components/responses/404_NotFound"
                    },
                    "401": {
                        "$ref": "#/components/responses/401_Unauthorized"
                    }
                }
            },
            "put": {
                "tags": [
                    "Motorinci Specification Items"
                ],
                "summary": "Memperbarui item spesifikasi motorinci yang sudah ada",
                "operationId": "updateMotorinciSpecificationItem",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Motorinci specification item ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Tinggi Tempat Duduk"
                                    },
                                    "unit": {
                                        "type": "string",
                                        "example": "mm"
                                    },
                                    "desc": {
                                        "type": "string",
                                        "example": "Jarak dari jok ke tanah."
                                    },
                                    "order": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/200_Success"
                    },
                    "404": {
                        "$ref": "#/components/responses/404_NotFound"
                    },
                    "422": {
                        "$ref": "#/components/responses/422_UnprocessableContent"
                    },
                    "401": {
                        "$ref": "#/components/responses/401_Unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/403_Forbidden"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Motorinci Specification Items"
                ],
                "summary": "Menghapus item spesifikasi motorinci",
                "operationId": "deleteMotorinciSpecificationItem",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Motorinci specification item ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "$ref": "#/components/responses/204_NoContent"
                    },
                    "404": {
                        "$ref": "#/components/responses/404_NotFound"
                    },
                    "401": {
                        "$ref": "#/components/responses/401_Unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/403_Forbidden"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "SuccessJsonResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Operasi berhasil"
                    },
                    "data": {
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ErrorJsonResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Terjadi kesalahan"
                    }
                },
                "type": "object"
            },
            "ValidationErrorJsonResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Data yang diberikan tidak valid."
                    },
                    "errors": {
                        "type": "object",
                        "example": {
                            "field_name": [
                                "Pesan error."
                            ]
                        }
                    }
                },
                "type": "object"
            }
        },
        "responses": {
            "200_Success": {
                "description": "Operasi berhasil dieksekusi.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/SuccessJsonResponse"
                        }
                    }
                }
            },
            "201_Created": {
                "description": "Sumber daya berhasil dibuat.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/SuccessJsonResponse"
                        }
                    }
                }
            },
            "204_NoContent": {
                "description": "Operasi berhasil tetapi tidak ada konten untuk dikembalikan."
            },
            "401_Unauthorized": {
                "description": "Autentikasi gagal. Token tidak valid atau tidak ada.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ErrorJsonResponse"
                        }
                    }
                }
            },
            "403_Forbidden": {
                "description": "Akses ditolak. Anda tidak memiliki izin.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ErrorJsonResponse"
                        }
                    }
                }
            },
            "404_NotFound": {
                "description": "Sumber daya yang diminta tidak ditemukan.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ErrorJsonResponse"
                        }
                    }
                }
            },
            "422_UnprocessableContent": {
                "description": "Validasi gagal.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ValidationErrorJsonResponse"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Masukkan token Bearer Anda. Contoh: 'Bearer 1|abcdef...'",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}